gossipd: use u64 for broadcast index.
uintmap uses uint64_t, we should here too.

Couldn't resist drive-by style fix for increment :)

+++ gossipd/broadcast.h
@@ -20,7 +20,7 @@ struct queued_message {
 };
 
 struct broadcast_state {
-	u32 next_index;
+	u64 next_index;
 	UINTMAP(struct queued_message *) broadcasts;
 };
